UT Moved RegisterExecuteUT to com.test.thalitest
